FEP-d767: Extend ActivityPub with Valueflows (#74)
This is the start of a proposal for a new FEP, to extend ActivityPub with Valueflows vocabulary.  It's more like an outline, there are a lot of open questions, and we'll be seeking feedback and advice from the fediverse community in SocialHub.  If it is good enough for a DRAFT, feel free to remove the WIP; otherwise we're happy to address issues first.

Co-authored-by: Lynn Foster <lynn@mikorizal.org>
Reviewed-on: https://codeberg.org/fediverse/fep/pulls/74
Co-authored-by: lynnfoster <lynnfoster@noreply.codeberg.org>
Co-committed-by: lynnfoster <lynnfoster@noreply.codeberg.org>