OAuth 2.0 Profile for the ActivityPub API (#162)
The ActivityPub specification does not provide specific guidance on using OAuth 2.0 with the ActivityPub API ("c2s").

There are some special requirements for an API standard (scopes, client ids, client metadata) that are different than APIs with a single implementation.

This OAuth 2.0 profile provides guidance for clients and servers to allow smooth interoperability for OAuth 2.0.

Co-authored-by: Evan Prodromou <codeberg@evanprodromou.name>
Reviewed-on: https://codeberg.org/fediverse/fep/pulls/162
Co-authored-by: Evan Prodromou <evanp@noreply.codeberg.org>
Co-committed-by: Evan Prodromou <evanp@noreply.codeberg.org>